+++ src/apps/toolbar/index.tsx
@@ -3,7 +3,7 @@ import { createRoot } from 'react-dom/client';
 import { I18nextProvider } from 'react-i18next';
 import { Provider } from 'react-redux';
 
-import { loadStore, registerStoreEvents, store } from './modules/shared/store/infra';
+import { registerStoreEvents, store } from './modules/shared/store/infra';
 import { loadConstants } from './modules/shared/utils/infra';
 
 import { App } from './app';
@@ -17,22 +17,17 @@ import './styles/variables.css';
 import '../shared/styles/reset.css';
 import './styles/global.css';
 
-async function Main() {
-  wrapConsole();
-  await declareDocumentAsLayeredHitbox();
-  await loadConstants();
-  await loadStore();
-  await registerStoreEvents();
-  await loadTranslations();
+wrapConsole();
+await declareDocumentAsLayeredHitbox();
+await loadConstants();
+await registerStoreEvents();
+await loadTranslations();
 
-  const container = getRootContainer();
-  createRoot(container).render(
-    <Provider store={store}>
-      <I18nextProvider i18n={i18n}>
-        <App />
-      </I18nextProvider>
-    </Provider>,
-  );
-}
-
-Main();
+const container = getRootContainer();
+createRoot(container).render(
+  <Provider store={store}>
+    <I18nextProvider i18n={i18n}>
+      <App />
+    </I18nextProvider>
+  </Provider>,
+);