+++ src/apps/toolbar/modules/shared/store/app.ts
@@ -1,5 +1,19 @@
 import { createSelector, createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { DocumentsFolder, DownloadsFolder, MusicFolder, PicturesFolder, RecentFolder, Settings, UIColors, VideosFolder } from '@seelen-ui/lib';
+import {
+  ApplicationHistory,
+  DesktopFolder,
+  DocumentsFolder,
+  DownloadsFolder,
+  invoke,
+  MusicFolder,
+  PicturesFolder,
+  RecentFolder,
+  SeelenCommand,
+  Settings,
+  UIColors,
+  UserDetails,
+  VideosFolder,
+} from '@seelen-ui/lib';
 import { Placeholder, ToolbarItem } from '@seelen-ui/lib/types';
 
 import { RootState } from './domain';
@@ -10,24 +24,21 @@ const settings = await Settings.default();
 
 const initialState: RootState = {
   version: 0,
-  items: {
-    left: [],
-    center: [],
-    right: [],
-  },
+  items: await invoke(SeelenCommand.StateGetToolbarItems),
   plugins: [],
   dateFormat: '',
   isOverlaped: false,
-  user: null,
-  userRecentFolder: RecentFolder.default().all(),
-  userDocumentsFolder: DocumentsFolder.default().all(),
-  userDownloadsFolder: DownloadsFolder.default().all(),
-  userPicturesFolder: PicturesFolder.default().all(),
-  userVideosFolder: VideosFolder.default().all(),
-  userMusicFolder: MusicFolder.default().all(),
+  user: (await UserDetails.getAsync()).user,
+  userRecentFolder: (await RecentFolder.getAsync()).asArray(),
+  userDesktopFolder: (await DesktopFolder.getAsync()).asArray(),
+  userDocumentsFolder: (await DocumentsFolder.getAsync()).asArray(),
+  userDownloadsFolder: (await DownloadsFolder.getAsync()).asArray(),
+  userPicturesFolder: (await PicturesFolder.getAsync()).asArray(),
+  userVideosFolder: (await VideosFolder.getAsync()).asArray(),
+  userMusicFolder: (await MusicFolder.getAsync()).asArray(),
   focused: null,
   settings: settings.fancyToolbar,
-  env: {},
+  env: (await invoke(SeelenCommand.GetUserEnvs)) as Record<string, string>,
   // default values of https://learn.microsoft.com/en-us/windows/win32/api/winbase/ns-winbase-system_power_status
   powerStatus: {
     acLineStatus: 255,
@@ -49,8 +60,8 @@ const initialState: RootState = {
   mediaOutputs: [],
   mediaInputs: [],
   notifications: [],
-  history: [],
-  historyOnMonitor: [],
+  history: (await ApplicationHistory.getAsync()).asArray(),
+  historyOnMonitor: (await ApplicationHistory.getCurrentMonitorHistoryAsync()).asArray(),
   colors: UIColors.default().inner,
 };