+++ changelog.md
@@ -7,6 +7,7 @@
 - toggle dock items using win + number.
 - notifications count on dock app items.
 - add brightness slider to quick settings on toolbar.
+- add user module on toolbar.
 
 ### refactor
 - create separated system service to handle elevated actions.